Updated command-line flags and plot view with bucket size.
This commit updates the command-line flags in the `cmd/run/main.go` file to include the bucket size option. The `plotView` function in the `cmd/ui/main.go` file has been modified to accept the bucket size as a parameter and use it to compute the cumulative distribution function (CDF). The CDF is then used to create plots of the received R and received R-1 values. Additionally, the `computeCDF` function has been updated to use the bucket size to group data points.